CAT-203 Assess and Assessment Wizard reogranisation and cleanup
