+++ src/components/index.ts
@@ -2,18 +2,5 @@ import { Header } from "./Header";
 import { Footer } from "./Footer";
 import { CustomTable } from "./CustomTable";
 import { Alert } from "./Alert";
-import { ActorCard } from "./ActorCard";
-import { EvidenceURLS } from "./EvidenceURLS";
-import { AssessmentInfo } from "./assessment/AssessmentInfo";
-import { AssessmentTabs } from "./assessment/AssessmentTabs";
 
-export {
-  Header,
-  CustomTable,
-  Footer,
-  Alert,
-  ActorCard,
-  EvidenceURLS,
-  AssessmentInfo,
-  AssessmentTabs,
-};
+export { Header, CustomTable, Footer, Alert };
