--- src/components/tests/index.ts
@@ -1,4 +1,0 @@
-import { TestBinaryForm } from "./TestBinaryForm";
-import { TestValueForm } from "./TestValueForm";
-
-export { TestBinaryForm, TestValueForm };
