+++ src/pages/assessments/components/index.ts
@@ -1,0 +1,4 @@
+export * from "./tests";
+export * from "./ActorCard";
+export * from "./AssessmentInfo";
+export * from "./CriteriaTabs";
