+++ src/pages/assessments/components/tests/index.ts
@@ -1,0 +1,5 @@
+import { TestBinaryForm } from "./TestBinaryForm";
+import { TestValueForm } from "./TestValueForm";
+import { EvidenceURLS } from "./EvidenceURLS";
+
+export { TestBinaryForm, TestValueForm, EvidenceURLS };
